# Reduce memory use of file_store.rb