@nateberkopec 2018-09-12
@schneems With benchmark/ips, I usually like to see error numbers (here ±12.2%) less than 5%. What happens if you run the benchmark for longer than the default 5 seconds?