@schneems 2018-09-12
Bumping up to 30 seconds still gives me a large error:

```
irb(main):018:0> Benchmark.ips do |x|
irb(main):019:1*   x.report("no-sync") { no_sync }
irb(main):020:1>   x.report("sync   ") { sync }
irb(main):021:1>   x.time = 30
irb(main):022:1>   x.compare!
irb(main):023:1> end
Warming up --------------------------------------
             no-sync   606.000  i/100ms
             sync      592.000  i/100ms
Calculating -------------------------------------
             no-sync      5.999k (±15.5%) i/s -    175.134k in  30.031027s
             sync         6.018k (±16.5%) i/s -    174.640k in  30.074512s

Comparison:
             sync   :     6017.8 i/s
             no-sync:     5999.3 i/s - same-ish: difference falls within error
```

Even with 100 seconds, it's a really high variance. I'm guessing the disk IO is the variance culprit.