@schneems 2018-09-12
It was pointed out that  the `file.write` method is faster for lots of data rather than manually passing in a file object to Marshal.dump btw. The PR was already using the `file.write` method so it doesn't need to be updated. Just mentioning for completeness. Thanks to .

```ruby
require 'benchmark/ips'

def rand_str
  (0...50).map { ('a'..'z').to_a[rand(26)] }.join
end
big_data = []
10_000.times { big_data << [rand_str, nil, rand(10_000) + rand() ] }

DATA = big_data

def no_write
  File.open("tmp-sync.dump", 'w+') do |f|
    f.sync = true
    Marshal.dump(DATA, f)
  end
end

def write
  File.open("tmp-no-sync.dump", 'w+') do |f|
    f.sync = true
    f.write Marshal.dump(DATA)
  end
end

Benchmark.ips do |x|
  x.report("no-write") { no_write }
  x.report("write   ") { write }
  x.compare!
end
# Warming up --------------------------------------
#             no-write     4.000  i/100ms
#             write        5.000  i/100ms
# Calculating -------------------------------------
#             no-write     43.150  (± 7.0%) i/s -    216.000  in   5.030772s
#             write        60.228  (± 1.7%) i/s -    305.000  in   5.067139s

# Comparison:
#             write   :       60.2 i/s
#             no-write:       43.1 i/s - 1.40x  slower
```