@Fryguy 2018-09-13
Both of these use File.open, where you have this control.  I'm curious; how does File.write stack up?