@Fryguy 2018-09-13
https://ruby-doc.org/core-2.4.2/IO.html#method-c-write

Was on mobile at the time I posted, so couldn't test myself... I'll give it a shot in a few hours when I'm back at a computer.