@Fryguy 2018-09-13
Tweaked your original script to do "wb+" mode as per the actual PR, and then added all of the modes (file.write, dump direct, and File.write)...

```ruby
require 'benchmark/ips'

DATA = ["1ada37891cbb5442fa56157bc451a608", nil, 634602.016795]

def sync_write
  File.open("tmp-sync.dump", 'wb+') do |f|
    f.sync = true
    f.write Marshal.dump(DATA)
  end
end

def no_sync_write
  File.open("tmp-no-sync.dump", 'wb+') do |f|
    f.write Marshal.dump(DATA)
  end
end

def sync_dump_direct
  File.open("tmp-sync.dump", 'wb+') do |f|
    f.sync = true
    Marshal.dump(DATA, f)
  end
end

def no_sync_dump_direct
  File.open("tmp-no-sync.dump", 'wb+') do |f|
    Marshal.dump(DATA, f)
  end
end

def file_write
  File.write("tmp-no-sync.dump", Marshal.dump(DATA), :mode => "wb+")
end

Benchmark.ips do |x|
  x.report("no-sync write       ") { no_sync_write }
  x.report("sync write          ") { sync_write }
  x.report("no-sync dump direct ") { no_sync_dump_direct }
  x.report("sync dump direct    ") { sync_dump_direct }
  x.report("File.write          ") { file_write }
  x.compare!
end

# Warming up --------------------------------------
# no-sync write          516.000  i/100ms
# sync write             505.000  i/100ms
# no-sync dump direct    495.000  i/100ms
# sync dump direct       493.000  i/100ms
# File.write             579.000  i/100ms
# Calculating -------------------------------------
# no-sync write             5.079k (± 6.8%) i/s -     25.284k in   5.003636s
# sync write                5.252k (± 7.9%) i/s -     26.260k in   5.035655s
# no-sync dump direct       5.006k (± 6.6%) i/s -     25.245k in   5.072092s
# sync dump direct          5.216k (± 6.5%) i/s -     26.129k in   5.032092s
# File.write                5.651k (± 6.6%) i/s -     28.371k in   5.045716s
#
# Comparison:
# File.write          :     5651.3 i/s
# sync write          :     5251.9 i/s - same-ish: difference falls within error
# sync dump direct    :     5216.2 i/s - same-ish: difference falls within error
# no-sync write       :     5078.8 i/s - same-ish: difference falls within error
# no-sync dump direct :     5005.9 i/s - same-ish: difference falls within error
```