@Fryguy 2018-09-13
Seems File.write is the "fastest", but the original purpose of this was for memory/allocations, and I'm not sure how you got those numbers.

Separate question, why "wb+" as opposed to "wb"?  The "+" makes it a read/write file handle, but since this is just dumping data, the "+" seems unnecessary.  I'm curious if removing the "+" helps out at all.