@schneems 2018-09-13
The "w+" is so if the file doesn't exist it gets created. It makes it easier to share the benchmark.